refactor: move usage to cli-usage
