+++ src/cli.js
@@ -7,6 +7,7 @@ const constant = require("./cli-constant");
 const util = require("./cli-util");
 const validator = require("./cli-validator");
 const normalizer = require("./cli-normalizer");
+const usage = require("./cli-usage");
 
 function run(args) {
   const argv = normalizer.normalizeArgv(
@@ -25,7 +26,7 @@ function run(args) {
   }
 
   if (argv["help"]) {
-    console.log(constant.usage);
+    console.log(usage);
     process.exit(0);
   }
 
@@ -39,7 +40,7 @@ function run(args) {
   } else if (hasFilePatterns) {
     util.formatFiles(argv);
   } else {
-    console.log(constant.usage);
+    console.log(usage);
     process.exit(1);
   }
 }
